tests/wpt/tests/navigation-api/scroll-behavior/scroll-after-preventDefault.html

<!doctype html>
<meta name="viewport" content="width=device-width,initial-scale=1,minimum-scale=1">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../navigation-methods/return-value/resources/helpers.js"></script>
<body>
<div style="height: 200vh; width: 200vw;"></div>
<div id="frag"></div>
<script>
promise_test(async t => {
  // Wait for after the load event so that the navigation doesn't get converted
  // into a replace navigation.
  await new Promise(resolve => window.onload = () => t.step_timeout(resolve, 0));

  navigation.addEventListener("navigate", t.step_func(e => {
    e.intercept();
    e.preventDefault();
    assert_throws_dom("InvalidStateError", () => e.scroll());
  }), { once : true });
  await assertBothRejectDOM(t, navigation.navigate("#frag"), "AbortError");
}, "scroll: scroll() should throw after preventDefault");
</script>
</body>
