LayoutTests/imported/w3c/web-platform-tests/navigation-api/scroll-behavior/scroll-after-preventDefault-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: AbortError: Navigation aborted

PASS scroll: scroll() should throw after preventDefault
